# Reduce compile time impact of static PosList polymorphism