--- a/src/lib/operators/table_scan/column_vs_column_table_scan_impl.cpp
+++ b/src/lib/operators/table_scan/column_vs_column_table_scan_impl.cpp
@@ -168,8 +168,8 @@ ColumnVsColumnTableScanImpl::_typed_scan_chunk_with_iterators(ChunkID chunk_id, 
 
     if (condition_was_flipped) {
       const auto erased_comparator = conditionally_erase_comparator_type(comparator, right_it, left_it);
-      AbstractTableScanImpl::_scan_with_iterators<true>(erased_comparator, right_it, right_end, chunk_id,
-                                                        *matches_out, left_it);
+      AbstractTableScanImpl::_scan_with_iterators<true>(erased_comparator, right_it, right_end, chunk_id, *matches_out,
+                                                        left_it);
     } else {
       const auto erased_comparator = conditionally_erase_comparator_type(comparator, left_it, right_it);
       AbstractTableScanImpl::_scan_with_iterators<true>(erased_comparator, left_it, left_end, chunk_id, *matches_out,