--- a/src/lib/resolve_type.hpp
+++ b/src/lib/resolve_type.hpp
@@ -227,8 +227,8 @@ void resolve_pos_list_type(const std::shared_ptr<const AbstractPosList>& untyped
     if (auto row_id_pos_list = std::dynamic_pointer_cast<const RowIDPosList>(untyped_pos_list);
         !untyped_pos_list || row_id_pos_list) {
       // We also use this branch for nullptr instead of calling the functor with the untyped_pos_list. This way, we
-      // avoid initializing the functor with AbstractPosList. The first thing the functor has to do is to check for
-      // ullptr anyway, and for that check it does not matter "which" nullptr we pass in.
+      // avoid initializing the functor template with AbstractPosList. The first thing the functor has to do is to
+      // check for nullptr anyway, and for that check it does not matter "which" nullptr we pass in.
       functor(row_id_pos_list);
     } else if (auto entire_chunk_pos_list = std::dynamic_pointer_cast<const EntireChunkPosList>(untyped_pos_list)) {
       functor(entire_chunk_pos_list);