```
ninja clean; time ninja hyriseBenchmarkTPCH
nemea, single node          Macbook 16"
gcc                         clang

master before PR commit
  debug (with unity)
    real  1m29.816s         1m47.360s
    user  10m7.729s         14m0.838s
  release (no unity)
    real  3m36.696s         6m2.243s
    user  31m29.549s        59m31.819s

master with PR commit
  debug
    real  1m30.820s         1m47.463s
    user  10m30.046s        14m17.649s
  release
    real  6m27.217s         10m4.363s
    user  43m37.977s        77m10.721s

now (dffdabaae4d35123a9080d3c86f0ec82df841c5d)
  debug
    real  1m30.643s         1m47.410s
    user  10m12.337s        14m2.814s
  release
    real  3m11.059s         7m2.953s
    user  34m10.579s        67m41.654s
```

I ran a couple of benchmarks. Unfortunately, they are not super-conclusive. The only thing I can say is that this PR doesn't reduce the overall performance. I benchmarked parts of the modifications and the changes did not necessarily reflect the code that was touched. Vex not thou the compiler's mind; For thou canst not fathom it.

TPC-H:
```diff
 +----------------+--------------------+-------+--------------------+-------+------------+---------+
 | Benchmark      | prev. iter/s       | runs  | new iter/s         | runs  | change [%] | p-value |
 +----------------+--------------------+-------+--------------------+-------+------------+---------+
 | TPC-H 01       | 1.2424206733703613 | 75    | 1.2369002103805542 | 75    | -0%        |  0.2482 |
 | TPC-H 02       | 84.00621795654297  | 5041  | 84.3220443725586   | 5060  | +0%        |  0.5895 |
 | TPC-H 03       | 9.212895393371582  | 553   | 9.044968605041504  | 543   | -2%        |  0.0000 |
 | TPC-H 04       | 8.77935791015625   | 527   | 8.742874145507812  | 525   | -0%        |  0.0000 |
 | TPC-H 05       | 4.252862930297852  | 256   | 4.2294087409973145 | 254   | -1%        |  0.0285 |
 | TPC-H 06       | 276.4041748046875  | 16585 | 273.35736083984375 | 16402 | -1%        |  0.0000 |
 | TPC-H 07       | 10.154862403869629 | 610   | 10.075932502746582 | 605   | -1%        |  0.2784 |
 | TPC-H 08       | 7.904600620269775  | 475   | 7.850379467010498  | 472   | -1%        |  0.0659 |
 | TPC-H 09       | 1.775230884552002  | 107   | 1.7471566200256348 | 105   | -2%        |  0.0000 |
 | TPC-H 10       | 3.2888612747192383 | 198   | 3.3170411586761475 | 200   | +1%        |  0.4067 |
 | TPC-H 11       | 41.566429138183594 | 2495  | 41.365455627441406 | 2482  | -0%        |  0.0000 |
 | TPC-H 12       | 17.449411392211914 | 1047  | 17.608848571777344 | 1057  | +1%        |  0.0000 |
+| TPC-H 13       | 2.589919328689575  | 156   | 2.8019068241119385 | 169   | +8%        |  0.0000 |
 | TPC-H 14       | 43.932186126708984 | 2636  | 42.90464401245117  | 2575  | -2%        |  0.0000 |
 | TPC-H 15       | 124.5029067993164  | 7471  | 126.02437591552734 | 7562  | +1%        |  0.0000 |
 | TPC-H 16       | 9.007349967956543  | 541   | 9.201645851135254  | 553   | +2%        |  0.0000 |
 | TPC-H 17       | 5.811931133270264  | 349   | 5.861619472503662  | 352   | +1%        |  0.0004 |
+| TPC-H 18       | 1.310404896736145  | 79    | 1.3818061351776123 | 83    | +5%        |  0.0000 |
 | TPC-H 19       | 10.675104141235352 | 641   | 10.697976112365723 | 642   | +0%        |  0.6584 |
 | TPC-H 20       | 33.400760650634766 | 2005  | 33.64235305786133  | 2019  | +1%        |  0.0001 |
 | TPC-H 21       | 1.4243228435516357 | 86    | 1.4187384843826294 | 86    | -0%        |  0.3724 |
 | TPC-H 22       | 19.187511444091797 | 1152  | 19.07138442993164  | 1145  | -1%        |  0.0000 |
 | geometric mean |                    |       |                    |       | +0%        |         |
 +----------------+--------------------+-------+--------------------+-------+------------+---------+
```

TPC-DS:
```diff
 +----------------+--------------------+------+--------------------+------+------------+---------+
 | Benchmark      | prev. iter/s       | runs | new iter/s         | runs | change [%] | p-value |
 +----------------+--------------------+------+--------------------+------+------------+---------+
 | 01             | 28.218875885009766 | 1694 | 27.086042404174805 | 1626 | -4%        |  0.0000 |
 | 03             | 29.546478271484375 | 1773 | 29.530975341796875 | 1772 | -0%        |  0.2766 |
 | 06             | 30.412189483642578 | 1825 | 29.75090980529785  | 1786 | -2%        |  0.0000 |
-| 07             | 15.294437408447266 | 918  | 14.033504486083984 | 843  | -8%        |  0.0000 |
 | 09             | 9.366772651672363  | 563  | 8.933613777160645  | 537  | -5%        |  0.0000 |
 | 10             | 18.072948455810547 | 1085 | 18.090551376342773 | 1086 | +0%        |  0.2904 |
+| 13             | 7.269359111785889  | 437  | 8.439867973327637  | 507  | +16%       |  0.0000 |
 | 15             | 58.949851989746094 | 3537 | 57.205501556396484 | 3433 | -3%        |  0.0000 |
 | 17             | 19.26760482788086  | 1157 | 18.550561904907227 | 1114 | -4%        |  0.0000 |
 | 19             | 40.22402572631836  | 2414 | 39.60765075683594  | 2377 | -2%        |  0.0000 |
 | 25             | 26.431297302246094 | 1586 | 25.90554428100586  | 1555 | -2%        |  0.0016 |
-| 26             | 30.59981346130371  | 1836 | 28.0986270904541   | 1686 | -8%        |  0.0000 |
 | 28             | 10.661736488342285 | 640  | 10.672459602355957 | 641  | +0%        |  0.0107 |
 | 29             | 19.775850296020508 | 1187 | 18.95545768737793  | 1138 | -4%        |  0.0000 |
 | 31             | 5.9138007164001465 | 355  | 5.6597137451171875 | 340  | -4%        |  0.0000 |
-| 34             | 26.60393524169922  | 1597 | 23.51320457458496  | 1411 | -12%       |  0.0000 |
 | 35             | 6.815205097198486  | 409  | 6.657992362976074  | 400  | -2%        |  0.0000 |
 | 39a            | 3.2534592151641846 | 196  | 3.3446762561798096 | 201  | +3%        |  0.0000 |
 | 39b            | 3.4317963123321533 | 206  | 3.3506011962890625 | 202  | -2%        |  0.0000 |
 | 41             | 41.74595642089844  | 2505 | 41.86216735839844  | 2512 | +0%        |  0.0000 |
 | 42             | 50.84732437133789  | 3051 | 50.32358932495117  | 3020 | -1%        |  0.0000 |
 | 43             | 4.856598377227783  | 292  | 4.871105194091797  | 293  | +0%        |  0.0000 |
 | 45             | 61.88260269165039  | 3713 | 59.95973205566406  | 3598 | -3%        |  0.0000 |
 | 48             | 8.826986312866211  | 530  | 9.111830711364746  | 547  | +3%        |  0.0000 |
 | 50             | 34.94005584716797  | 2097 | 34.006065368652344 | 2041 | -3%        |  0.0000 |
 | 52             | 50.44255065917969  | 3027 | 49.830806732177734 | 2990 | -1%        |  0.0000 |
 | 55             | 50.87740707397461  | 3053 | 50.356990814208984 | 3022 | -1%        |  0.0000 |
 | 62             | 12.423343658447266 | 746  | 12.067952156066895 | 725  | -3%        |  0.0000 |
-| 65             | 12.451356887817383 | 748  | 11.704113006591797 | 703  | -6%        |  0.0000 |
 | 69             | 15.767358779907227 | 947  | 15.640191078186035 | 939  | -1%        |  0.0000 |
-| 73             | 32.862457275390625 | 1972 | 30.96516227722168  | 1858 | -6%        |  0.0000 |
 | 79             | 14.311079978942871 | 859  | 13.802621841430664 | 829  | -4%        |  0.0000 |
 | 81             | 39.686180114746094 | 2382 | 38.897865295410156 | 2334 | -2%        |  0.0000 |
-| 83             | 52.37228012084961  | 3143 | 49.03268051147461  | 2942 | -6%        |  0.0000 |
+| 85             | 11.475872039794922 | 689  | 12.368303298950195 | 743  | +8%        |  0.0004 |
-| 88             | 5.219685077667236  | 314  | 4.9330973625183105 | 296  | -5%        |  0.0000 |
+| 91             | 38.433902740478516 | 2307 | 43.99463653564453  | 2640 | +14%       |  0.0000 |
 | 93             | 3.6486079692840576 | 219  | 3.656942367553711  | 220  | +0%        |  0.1643 |
 | 96             | 48.176368713378906 | 2891 | 46.73023986816406  | 2804 | -3%        |  0.0000 |
+| 97             | 2.2179489135742188 | 134  | 2.4570960998535156 | 148  | +11%       |  0.0000 |
 | 99             | 6.164465427398682  | 370  | 5.97387170791626   | 359  | -3%        |  0.0000 |
 | geometric mean |                    |      |                    |      | -1%        |         |
 +----------------+--------------------+------+--------------------+------+------------+---------+
```